Use the resolved tmux commands instead of tmux-activate-pane in the tmujx-activate-on-fail aliases
The pane activation also doesn't work from inside my rvtt; probably, the same problem with pipelining.
So just resolve the actual tmux commands for activation and directly embed that inside the alias. As long as that shell only is visible in a single session and pane, that should not make any difference.