tmux-connect: Adapt: Just show sessions again
In tmux 2.1, the tree is collapsed and sessions must be expanded, whereas in tmux 3.0a, the tree is expanded and -s must be given to start with sessions collapsed.
Fortunately, in the older tmux the -s argument will restrict the output to sessions (what the dropped choose-session is an alias of), so it will do fine there as well.