ENH: Indicate inactive COMMAND by wrapping in (...)
We need to pass the commandName to the new Tmux window (via -e), and obtain the current window ID so that we'll find ourselves again after COMMAND has finished.
As the current Tmux window may have already changed by the time command-launcher.sh is running, we check by comparing the passed commandName with the current one.