queried_command-launcher: ENH: Allow setting the window name via optional flag
This is used by my new prefix + * mapping that is a variant of prefix + - / prefix + |, using a new window instead of a split pane.
Like prefix + c, we want to use the command's name as the window name, and indicate a concluded command by putting its name in round parentheses, just like new-window-launcher.sh does.