Reimplementation: Use new-window-launcher.sh for the queried command
Instead of extending queried_command-launcher.sh with the optional window handling (which, with updating the window name and restoring it and the status is quite involved), this functionality is better combined with new-window-launcher.sh, which already contains a simpler previous implementation of the "prompt when the command's runtime was short" feature, and the "tmux SHELL-COMMAND" feature would benefit from that more elaborate implementation with ensurePrompting and runWithPrompt, too.
The only difference is in how the command is passed; with the tmux-wrapper, commandName is already resolved in the wrapper function, directly sets the Tmux window name (an inconsequential optimization that tries to avoid one tmux invocation), and passes the command. It's no big deal to change that to set the Tmux @queried_command variable, too, and with that the SHELL-COMMAND can now also be repeated through my prefix + g* / prefix + g- mappings.

FIX: Looking up windowId twice in new-window-launcher.sh looks questionable, and I think the lookup of windowName actually has to happen _after_ COMMAND has been executed, as the purpose of the conditionals is to avoid changing an explicitly renamed Tmux window name. All of that is now handled through the custom prompt command passed to ensurePrompting.