BUG: "error connecting to /tmp/tmux-4646/default (No such file or directory)" when tssh'ing
We need to differentiate between having Tmux available on the system (what the tmux-extension project is about), and running under (local) Tmux control (i.e. $TMUX is set). The hooks and MOTD need to be triggered by the latter. As running under local Tmux control means that Tmux needs to be installed, that's another correct part of the condition.
When SSH'ing into a remote host from a local SSH session, $TMUX is empty (as long as no nested Tmux session is initiated on the remote system).