Robustness: Handle closing of the split pane triggered from elsewhere
If the shell-command ends while another pane is active, the "tmux capture-pane" will capture the wrong (active) pane's contents!
If my withLess wrapper is used, this should normally not happen, as the shell-command is wrapped by ensurePrompting, and in order to acknowledge that, one needs to activate that pane again in order to register the keypress there.
However, ensurePrompting does not require that keypress if it detects that the pager has been used (to avoid having to quit twice). So in the case of active paging but that paging getting killed externally, the bug occurs.
This happened with "make up" of a broken container that quit almost immediately. The container startup is covered by my smartless, less is detected by ensurePrompting and it deactivates its own confirmation, smartless realizes the output fully fits the terminal and kills the pager. In order to shut down the container, a parallel "make down" needs to be issued, which I do via prefix + -, and that runs that command in another temporary split pane. Through that, the "make up" command exits, pager is gone already, and as the "make down" is active, its output is captured instead!

Fortunately, "tmux capture-pane" supports -t target-pane; it's just a bit of a hassle to correctly interpolate that into the trap command (to avoid an intermediate shell variable that would be visible to the executed shell-command). Interpolation has to happen during trap definition; at that point in time, the split pane still is the active one, and we can interpolate its unique ID to later correctly capture that, regardless of which pane is the active one then.