ENH: Allow match inversion via -v|--invert-match
It can be useful to show PIDs of other panes, e.g. when a process in another pane / window is handing, and we want to kill it. (I've accidentally had a Bash completion function waiting for input from readarray, and neither Ctrl-C nor EOF stopped it.)

System-info: Armbian 24.11.1 noble, aarch64
Platform-info: tmux 3.4